+++ tests/integration/test_logging_flush.py
@@ -13,8 +13,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License
 
-from multiprocessing.pool import ThreadPool
-
 from six.moves import mock
 
 from tests import REPORT_PORTAL_SERVICE
@@ -29,15 +27,8 @@ def test_logging_flushing(mock_client_init):
     """
     mock_client = mock_client_init.return_value
 
-    def run_test():
-        from reportportal_client._local import set_current
-        set_current(mock_client)
-        return utils.run_pytest_tests(['examples/test_rp_logging.py'])
-
-    pool = ThreadPool(processes=1)
-    async_result = pool.apply_async(run_test)
-    result = async_result.get()
-    pool.terminate()
+    result = utils.run_tests_with_client(
+        mock_client, ['examples/test_rp_logging.py'])
 
     assert int(result) == 0, 'Exit code should be 0 (no errors)'
     assert mock_client.terminate.call_count == 1, \